+++ packages/core/src/AnimatedFlowLinesLayer/AnimatedFlowLinesLayerVertex.glsl.ts
@@ -42,59 +42,72 @@ export default `\
 #define SHADER_NAME animated-flow-lines-layer-vertex-shader
 #define SPEED 0.015
 #define NUM_PARTS 5.0
-#define DIR_OFFSET 0.0 
 
 attribute vec3 positions;
 attribute vec3 instanceSourcePositions;
 attribute vec3 instanceTargetPositions;
-attribute vec4 instanceSourceTargetPositions64xyLow;
+attribute vec3 instanceSourcePositions64Low;
+attribute vec3 instanceTargetPositions64Low;
 attribute vec4 instanceColors;
 attribute vec3 instancePickingColors;
 attribute float instanceWidths;
-attribute float instanceStaggering;
 attribute float instancePickable;
-         
+attribute float instanceStaggering;
+
 uniform float opacity;
 uniform float currentTime;
 uniform float thicknessUnit;
     
 varying vec4 vColor;
 varying float sourceToTarget;
+varying vec2 uv;
 
 // offset vector by strokeWidth pixels
 // offset_direction is -1 (left) or 1 (right)
-vec2 getExtrusionOffset(vec2 line_clipspace, float offset_direction) {
+vec2 getExtrusionOffset(vec2 line_clipspace, float offset_direction, float width) {
   // normalized direction of the line
   vec2 dir_screenspace = normalize(line_clipspace * project_uViewportSize);
   // rotate by 90 degrees
   dir_screenspace = vec2(-dir_screenspace.y, dir_screenspace.x);
 
-  vec2 offset_screenspace = dir_screenspace * offset_direction * instanceWidths * thicknessUnit;
-  vec2 offset_clipspace = project_pixel_size_to_clipspace(offset_screenspace).xy;
-
-  return offset_clipspace;
+  return dir_screenspace * offset_direction * width / 2.0;
 }
 
 void main(void) {
-  // Position
-  vec4 source = project_position_to_clipspace(instanceSourcePositions, vec2(0.), vec3(0.));
-  vec4 target = project_position_to_clipspace(instanceTargetPositions, vec2(0.), vec3(0.));
+  geometry.worldPosition = instanceSourcePositions;
+  geometry.worldPositionAlt = instanceTargetPositions;
 
+  // Position
+  vec4 source_commonspace;
+  vec4 target_commonspace;
+  vec4 source = project_position_to_clipspace(instanceSourcePositions, instanceSourcePositions64Low, vec3(0.), source_commonspace);
+  vec4 target = project_position_to_clipspace(instanceTargetPositions, instanceTargetPositions64Low, vec3(0.), target_commonspace);
+
+  float widthPixels = instanceWidths * thicknessUnit;
+  
+  
   // linear interpolation of source & target to pick right coord
   float segmentIndex = positions.x;
   vec4 p = mix(source, target, segmentIndex);
-
+  geometry.position = mix(source_commonspace, target_commonspace, segmentIndex);
+  uv = positions.xy;
+  geometry.uv = uv;
+  if (instancePickable > 0.5) {
+    geometry.pickingColor = instancePickingColors;
+  }
+  
   // extrude
-  vec2 offset = getExtrusionOffset(target.xy - source.xy, positions.y - DIR_OFFSET);
-  gl_Position = p + vec4(offset, 0.0, 0.0);
+  vec3 offset = vec3(
+    getExtrusionOffset(target.xy - source.xy, positions.y, widthPixels),
+    0.0);
+  DECKGL_FILTER_SIZE(offset, geometry);
+  gl_Position = p + vec4(project_pixel_size_to_clipspace(offset.xy), 0.0, 0.0);
+  DECKGL_FILTER_GL_POSITION(gl_Position, geometry);
 
   // Color
   vColor = vec4(instanceColors.rgb, instanceColors.a * opacity) / 255.;
-  sourceToTarget = positions.x * length(source - target) * NUM_PARTS - currentTime * SPEED + instanceStaggering; 
+  DECKGL_FILTER_COLOR(vColor, geometry);
 
-  // Set color to be rendered to picking fbo (also used to check for selection highlight).
-  if (instancePickable > 0.5) {
-    picking_setPickingColor(instancePickingColors);
-  }
+  sourceToTarget = positions.x * length(source - target) * NUM_PARTS - currentTime * SPEED + instanceStaggering; 
 }
 `;