+++ packages/core/src/FlowLinesLayer/FlowLinesLayerFragment.glsl.ts
@@ -40,23 +40,18 @@
 export default `\
 #define SHADER_NAME flow-line-layer-fragment-shader
 
-#ifdef GL_ES
 precision highp float;
-#endif
 
 varying vec4 vColor;
+varying vec2 uv;
 
 void main(void) {
   if (vColor.a == 0.0) {
     discard;
   }
 
+  geometry.uv = uv;
   gl_FragColor = vColor;
-  
-  // use highlight color if this fragment belongs to the selected object.
-  gl_FragColor = picking_filterHighlightColor(gl_FragColor);
-  
-  // use picking color if rendering to picking FBO.
-  gl_FragColor = picking_filterPickingColor(gl_FragColor);
+  DECKGL_FILTER_COLOR(gl_FragColor, geometry);
 }
 `;