+++ packages/core/src/Selectors.ts
@@ -18,7 +18,7 @@ import {
   isDiffColorsRGBA,
   RGBA,
 } from './colors';
-import { Props } from './FlowMapLayer';
+import FlowMapLayer, { Props } from './FlowMapLayer';
 import {
   Flow,
   FlowAccessors,
@@ -59,6 +59,8 @@ const getHighlightedFlow = (props: Props) => props.highlightedFlow;
 const getHighlightedLocationId = (props: Props) => props.highlightedLocationId;
 const getSelectedLocationIds = (props: Props) => props.selectedLocationIds;
 const getShowOnlyTopFlows = (props: Props) => props.showOnlyTopFlows;
+const getMaxLocationCircleSize = (props: Props) =>
+  props.maxLocationCircleSize != null ? props.maxLocationCircleSize : FlowMapLayer.defaultProps.maxLocationCircleSize;
 
 class Selectors {
   constructor(private inputAccessors: InputAccessors) {}
@@ -73,22 +75,19 @@ class Selectors {
     },
   );
 
-  getLocationByIdGetter: PropsSelector<LocationByIdGetter> = createSelector(
-    [getLocationFeatures],
-    locations => {
-      const locationsById = nest<Location, Location | undefined>()
-        .key(this.inputAccessors.getLocationId)
-        .rollup(([d]) => d)
-        .object(locations);
-      return (id: string) => {
-        const location = locationsById[id];
-        if (!location) {
-          console.warn(`No location found for id '${id}'`);
-        }
-        return location;
-      };
-    },
-  );
+  getLocationByIdGetter: PropsSelector<LocationByIdGetter> = createSelector([getLocationFeatures], locations => {
+    const locationsById = nest<Location, Location | undefined>()
+      .key(this.inputAccessors.getLocationId)
+      .rollup(([d]) => d)
+      .object(locations);
+    return (id: string) => {
+      const location = locationsById[id];
+      if (!location) {
+        console.warn(`No location found for id '${id}'`);
+      }
+      return location;
+    };
+  });
 
   private getFilteredFlows: PropsSelector<Flow[]> = createSelector(
     [getFlows, getSelectedLocationIds],
@@ -107,22 +106,16 @@ class Selectors {
     },
   );
 
-  private getNonSelfFlows: PropsSelector<Flow[]> = createSelector(
-    [this.getFilteredFlows],
-    flows => {
-      const { getFlowOriginId, getFlowDestId } = this.inputAccessors;
-      return flows.filter(flow => getFlowOriginId(flow) !== getFlowDestId(flow));
-    },
-  );
+  private getNonSelfFlows: PropsSelector<Flow[]> = createSelector([this.getFilteredFlows], flows => {
+    const { getFlowOriginId, getFlowDestId } = this.inputAccessors;
+    return flows.filter(flow => getFlowOriginId(flow) !== getFlowDestId(flow));
+  });
 
-  getSortedNonSelfFlows: PropsSelector<Flow[]> = createSelector(
-    [this.getNonSelfFlows],
-    flows => {
-      const comparator = (f1: Flow, f2: Flow) =>
-        Math.abs(this.inputAccessors.getFlowMagnitude(f1)) - Math.abs(this.inputAccessors.getFlowMagnitude(f2));
-      return flows.slice().sort(comparator);
-    },
-  );
+  getSortedNonSelfFlows: PropsSelector<Flow[]> = createSelector([this.getNonSelfFlows], flows => {
+    const comparator = (f1: Flow, f2: Flow) =>
+      Math.abs(this.inputAccessors.getFlowMagnitude(f1)) - Math.abs(this.inputAccessors.getFlowMagnitude(f2));
+    return flows.slice().sort(comparator);
+  });
 
   getTopFlows: PropsSelector<Flow[]> = createSelector(
     [this.getSortedNonSelfFlows, getShowOnlyTopFlows],
@@ -155,7 +148,7 @@ class Selectors {
 
   private getFlowMagnitudeExtent: PropsSelector<[number, number] | [undefined, undefined]> = createSelector(
     [this.getNonSelfFlows],
-    flows => extent(flows, this.inputAccessors.getFlowMagnitude),
+    flows => extent(flows, f => this.inputAccessors.getFlowMagnitude(f)),
   );
 
   getFlowThicknessScale: PropsSelector<NumberScale> = createSelector(
@@ -336,12 +329,12 @@ class Selectors {
   );
 
   private getSizeScale: PropsSelector<NumberScale> = createSelector(
-    [this.getMaxLocationMaxAbsTotal],
-    maxTotal => {
+    [getMaxLocationCircleSize, this.getMaxLocationMaxAbsTotal],
+    (maxLocationCircleSize, maxTotal) => {
       const scale = scalePow()
         .exponent(1 / 2)
         .domain([0, maxTotal])
-        .range([0, maxTotal > 0 ? 15 : 1]);
+        .range([0, maxTotal > 0 ? maxLocationCircleSize : 1]);
 
       return (v: number) => scale(Math.abs(v));
     },